Settings sialog: code adjusments: warnings, cleanup, code style

+++ src/MEGASync/gui/StreamingFromMegaDialog.cpp
@@ -331,8 +331,8 @@ void StreamingFromMegaDialog::requestPublicNodeInfo()
 void StreamingFromMegaDialog::updateFileInfo(QString fileName, LinkStatus status)
 {
     ui->lFileName->ensurePolished();
-    ui->lFileName->setText(ui->lFileName->fontMetrics().elidedText(fileName, Qt::ElideMiddle,ui->lFileName->maximumWidth()));
-    ui->lFileSize->setText(Utilities::getSizeString(selectedMegaNode->getSize()));
+    ui->lFileName->setText(ui->lFileName->fontMetrics().elidedText(fileName,Qt::ElideMiddle,ui->lFileName->maximumWidth()));
+    ui->lFileSize->setText(Utilities::getSizeString(static_cast<unsigned long long>(selectedMegaNode->getSize())));
 
     QIcon typeIcon = Utilities::getExtensionPixmapMedium(fileName);
